add loading data stats summary from store

+++ src/app/services/http-client.service.ts
@@ -11,25 +11,28 @@ export class HttpClientService {
   private _apiRootUrl: string;
   private _systemInfo: string;
   private httpOptions: any;
+  private username = 'portalobservatory';
+  private password = 'DHIS2017';
+  private token: string;
   constructor(private httpClient: HttpClient, private manifestService: ManifestService) {
-    // this.token = btoa(this.username + ':' + this.password);
+    this.token = btoa(this.username + ':' + this.password);
   }
 
 
   get(url: string, preferPreviousApiVersion: boolean = false, useRootUrl: boolean = false): Observable<any> {
     const rootUrlPromise = useRootUrl ? this._getRootUrl() : this._getApiRootUrl(preferPreviousApiVersion);
     return rootUrlPromise.pipe(
-      mergeMap(rootUrl => this.httpClient.get(rootUrl + url).pipe(catchError(this._handleError)))
+      mergeMap(rootUrl => this.httpClient.get(rootUrl + url, {headers: { 'Authorization': 'Basic ' + this.token, 'Content-Type': 'application/json'}}).pipe(catchError(this._handleError)))
     );
   }
 
   post(url: string, data: any, preferPreviousApiVersion: boolean = false, useRootUrl: boolean = false,
-    headerOptions?: any) {
+       headerOptions?: any) {
     const rootUrlPromise = useRootUrl ? this._getRootUrl() : this._getApiRootUrl(preferPreviousApiVersion);
     return rootUrlPromise.pipe(
       mergeMap(rootUrl =>
-        this.httpClient.post(rootUrl + url, data).
-          pipe(catchError(this._handleError))
+        this.httpClient.post(rootUrl + url, data, {headers: { 'Authorization': 'Basic ' + this.token, 'Content-Type': 'application/json'}}).
+        pipe(catchError(this._handleError))
       )
     );
   }
@@ -38,7 +41,7 @@ export class HttpClientService {
     const rootUrlPromise = useRootUrl ? this._getRootUrl() : this._getApiRootUrl(preferPreviousApiVersion);
     return rootUrlPromise.pipe(
       mergeMap(rootUrl =>
-        this.httpClient.put(rootUrl + url, data).pipe(catchError(this._handleError))
+        this.httpClient.put(rootUrl + url, data, {headers: { 'Authorization': 'Basic ' + this.token, 'Content-Type': 'application/json'}}).pipe(catchError(this._handleError))
       )
     );
   }
@@ -46,14 +49,14 @@ export class HttpClientService {
   delete(url: string, preferPreviousApiVersion: boolean = false, useRootUrl: boolean = false) {
     const rootUrlPromise = useRootUrl ? this._getRootUrl() : this._getApiRootUrl(preferPreviousApiVersion);
     return rootUrlPromise.pipe(
-      mergeMap(rootUrl => this.httpClient.delete(rootUrl + url).pipe(catchError(this._handleError)))
+      mergeMap(rootUrl => this.httpClient.delete(rootUrl + url, {headers: { 'Authorization': 'Basic ' + this.token, 'Content-Type': 'application/json'}}).pipe(catchError(this._handleError)))
     );
   }
 
   getSystemInfo() {
     return this._systemInfo ? of(this._systemInfo) :
-      this._getRootUrl().pipe(switchMap((rootUrl: string) => this.httpClient.get(`${rootUrl}api/system/info`).
-        pipe(tap((systemInfo: any) => this._systemInfo = systemInfo))));
+      this._getRootUrl().pipe(switchMap((rootUrl: string) => this.httpClient.get(`${rootUrl}api/system/info`, {headers: { 'Authorization': 'Basic ' + this.token, 'Content-Type': 'application/json'}}).
+      pipe(tap((systemInfo: any) => this._systemInfo = systemInfo))));
   }
 
   // Private methods
@@ -101,12 +104,12 @@ export class HttpClientService {
 
   private _getApiRootUrl(preferPreviousVersion: boolean = false) {
     const rootUrlPromise = this._getRootUrl().
-      pipe(switchMap((rootUrl) => this.getSystemInfo().pipe(map((systemInfo) => {
-          const splitedVersion = systemInfo.version.split('.');
-          const version = parseInt(splitedVersion[1], 10);
-          return {rootUrl, version: (version - 1) <= 25 ? (version + 1) : version};
-        }))
-      ));
+    pipe(switchMap((rootUrl) => this.getSystemInfo().pipe(map((systemInfo) => {
+        const splitedVersion = systemInfo.version.split('.');
+        const version = parseInt(splitedVersion[1], 10);
+        return {rootUrl, version: (version - 1) <= 25 ? (version + 1) : version};
+      }))
+    ));
     return rootUrlPromise.pipe(
       map((urlInfo: {rootUrl: string, version: number}) => `${urlInfo.rootUrl}api/${preferPreviousVersion ?
         urlInfo.version ? ((urlInfo.version - 1) + '/') : '' : ''}`));
